% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal_portfolio.R
\name{optimal_portfolio}
\alias{optimal_portfolio}
\title{Generate Optimal Portfolio}
\usage{
optimal_portfolio(
  returns_data,
  target_return = NULL,
  up_weight = NULL,
  down_weight = NULL,
  split = 0,
  verbose = TRUE
)
}
\arguments{
\item{returns_data}{(dataframe): Input dataframe for which optimal portfolio is selected from.}

\item{target_return}{(numeric/vector): Optional parameter that constraints optimal portfolio to have specific return.
Default is NULL.}

\item{up_weight}{(numeric): Optional parameter that sets upside weight in asymmetric risk. Default is NULL,
which equates to a symmetric portfolio.}

\item{down_weight}{(numeric): Optional parameter that sets downside weight in asymmetric risk. Default is NULL,
which equates to a symmetric portfolio.}

\item{split}{(numeric/character): Optional parameter that is threshold that determines where to split data between upside
and downside risk. Can be a numeric value or 'mean'/'median'. Default is 0.}

\item{verbose}{(boolean): Print informative messages about removed columns.}
}
\value{
(list) weights (vector): Optimal weights selected to minimize portfolio standard deviations.,
minimum_risk (numeric): Variance of the optimal portfolio.,
expected_return (numeric): Expected return of optimal portfolio, given optimal weights
and mean returns of each revenue category.
}
\description{
This function selects the optimal municipal revenue portfolio based on parameter
values given (Optionally with semivariance weighting).
}
